split obsoletes by version
